remove useless entity
