--- hw6-books-jpa/src/main/java/ru/otus/dao/BookCommentsDao.java
@@ -1,7 +1,0 @@
-package ru.otus.dao;
-
-import ru.otus.domain.BookComment;
-
-public interface BookCommentsDao {
-	void save(BookComment comment);
-}
