+++ hw6-books-jpa/src/main/java/ru/otus/domain/Comment.java
@@ -6,12 +6,12 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 @Entity
-@Table(name = "COMMENTS")
+@Table(name = "BOOKS_COMMENTS")
 public class Comment {
 	@Id
 	@Column(name = "COMMENT_ID")
-	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "comment_s")
-	@SequenceGenerator(name = "comment_s", sequenceName = "COMMENTS_S", allocationSize = 1)
+	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "book_comment_s")
+	@SequenceGenerator(name = "book_comment_s", sequenceName = "BOOK_COMMENTS_S", allocationSize = 1)
 	private Long id;
 
 	@NotNull
@@ -20,12 +20,18 @@ public class Comment {
 	@Column(name = "TEXT", nullable = false, length = 255)
 	private String text;
 
+	@NotNull
+	@JoinColumn(name = "BOOK_ID", nullable = false)
+	@ManyToOne(targetEntity = Book.class, cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+	private Book book;
+
 	public Comment() {
 	}
 
 	private Comment(final Builder builder) {
 		this.id = builder.id;
 		this.text = builder.text;
+		this.book = builder.book;
 	}
 
 	public Long getId() {
@@ -44,17 +50,27 @@ public class Comment {
 		this.text = text;
 	}
 
+	public Book getBook() {
+		return book;
+	}
+
+	public void setBook(final Book book) {
+		this.book = book;
+	}
+
 	@Override
 	public String toString() {
 		return "Comment{" +
 				"id=" + id +
 				", text='" + text + '\'' +
+				", book=" + book +
 				'}';
 	}
 
 	public static class Builder {
 		private Long id;
 		private String text;
+		private Book book;
 
 		public Builder id(long id) {
 			this.id = id;
@@ -63,6 +79,11 @@ public class Comment {
 
 		public Builder text(String text) {
 			this.text = text;
+			return this;
+		}
+
+		public Builder book(Book book) {
+			this.book = book;
 			return this;
 		}
