--- hw6-books-jpa/src/main/java/ru/otus/services/BookCommentsService.java
@@ -1,7 +1,0 @@
-package ru.otus.services;
-
-import ru.otus.domain.BookComment;
-
-public interface BookCommentsService {
-	void save(BookComment bookComment);
-}
