--- hw6-books-jpa/src/main/java/ru/otus/services/facades/BookCommentsFacade.java
@@ -1,7 +1,0 @@
-package ru.otus.services.facades;
-
-import ru.otus.domain.Comment;
-
-public interface BookCommentsFacade {
-	void save(long bookId, Comment comment);
-}
